+++ Assets/OxGFrame/AssetLoader/ThirdParty/SimpleDiskUtils/Runtime/DiskUtils.cs
@@ -141,7 +141,7 @@ namespace SimpleDiskUtils
 	    /// </summary>
 	    /// <returns>The available space in MB.</returns>
 	    public static int CheckAvailableSpace(){
-		    DriveInfo drive = GetDrive (DEFAULT_DRIVE);
+		    DriveInfo drive = new DriveInfo(DEFAULT_DRIVE);
 		    if (drive == null)
 			    return -1;
 		    return int.Parse((drive.AvailableFreeSpace / MEGA_BYTE).ToString());
@@ -152,7 +152,7 @@ namespace SimpleDiskUtils
 	    /// </summary>
 	    /// <returns>The total space in MB.</returns>
 	    public static int CheckTotalSpace(){
-		    DriveInfo drive = GetDrive (DEFAULT_DRIVE);
+		    DriveInfo drive = new DriveInfo(DEFAULT_DRIVE);
 		    if (drive == null)
 			    return -1;
 		    return int.Parse ((drive.TotalSize / MEGA_BYTE).ToString());
@@ -163,7 +163,7 @@ namespace SimpleDiskUtils
 	    /// </summary>
 	    /// <returns>The busy space in MB.</returns>
 	    public static int CheckBusySpace(){
-		    DriveInfo drive = GetDrive (DEFAULT_DRIVE);
+		    DriveInfo drive = new DriveInfo(DEFAULT_DRIVE);
 		    if (drive == null)
 			    return -1;