+++ Assets/OxGFrame/AssetLoader/ThirdParty/YooAsset/Runtime/CacheSystem/PersistentHelper.cs
@@ -15,7 +15,7 @@ namespace YooAsset
 #elif UNITY_IPHONE
             return StringUtility.Format("file://{0}", path);
 #elif UNITY_ANDROID
-            if (path.StartsWith("jar:file//"))
+            if (path.StartsWith("jar:file://"))
                 return path;
             else
                 return StringUtility.Format("jar:file://{0}", path);
@@ -23,6 +23,8 @@ namespace YooAsset
             return new System.Uri(path).ToString();
 #elif UNITY_STANDALONE
             return StringUtility.Format("file:///{0}", path);
+#elif UNITY_OPENHARMONY
+            return path;
 #else
             return path;
 #endif