+++ Assets/OxGFrame/AssetLoader/ThirdParty/YooAsset/Runtime/ResourceManager/Provider/BundledSceneProvider.cs
@@ -9,14 +9,14 @@ namespace YooAsset
     internal sealed class BundledSceneProvider : ProviderBase
     {
         public readonly LoadSceneMode SceneMode;
-        private readonly bool _suspendLoad;
         private AsyncOperation _asyncOperation;
+        private bool _suspendLoadMode;
 
         public BundledSceneProvider(ResourceManager manager, string providerGUID, AssetInfo assetInfo, LoadSceneMode sceneMode, bool suspendLoad) : base(manager, providerGUID, assetInfo)
         {
             SceneMode = sceneMode;
             SceneName = Path.GetFileNameWithoutExtension(assetInfo.AssetPath);
-            _suspendLoad = suspendLoad;
+            _suspendLoadMode = suspendLoad;
         }
         internal override void InternalOnStart()
         {
@@ -35,6 +35,12 @@ namespace YooAsset
             // 1. 检测资源包
             if (_steps == ESteps.CheckBundle)
             {
+                if (IsWaitForAsyncComplete)
+                {
+                    DependBundles.WaitForAsyncComplete();
+                    OwnerBundle.WaitForAsyncComplete();
+                }
+
                 if (DependBundles.IsDone() == false)
                     return;
                 if (OwnerBundle.IsDone() == false)
@@ -60,39 +66,68 @@ namespace YooAsset
             // 2. 加载场景
             if (_steps == ESteps.Loading)
             {
-                // 注意：如果场景不存在则返回NULL
-                _asyncOperation = SceneManager.LoadSceneAsync(MainAssetInfo.AssetPath, SceneMode);
-                if (_asyncOperation != null)
+                if (IsWaitForAsyncComplete || IsForceDestroyComplete)
                 {
-                    _asyncOperation.allowSceneActivation = !_suspendLoad;
-                    _asyncOperation.priority = 100;
-                    SceneObject = SceneManager.GetSceneAt(SceneManager.sceneCount - 1);
+                    // 注意：场景同步加载方法不会立即加载场景，而是在下一帧加载。
+                    LoadSceneParameters parameters = new LoadSceneParameters(SceneMode);
+                    SceneObject = SceneManager.LoadScene(MainAssetInfo.AssetPath, parameters);
                     _steps = ESteps.Checking;
                 }
                 else
                 {
-                    string error = $"Failed to load scene : {MainAssetInfo.AssetPath}";
-                    YooLogger.Error(error);
-                    InvokeCompletion(error, EOperationStatus.Failed);
+                    // 注意：如果场景不存在异步加载方法返回NULL
+                    // 注意：即使是异步加载也要在当帧获取到场景对象
+                    _asyncOperation = SceneManager.LoadSceneAsync(MainAssetInfo.AssetPath, SceneMode);
+                    if (_asyncOperation != null)
+                    {
+                        _asyncOperation.allowSceneActivation = !_suspendLoadMode;
+                        _asyncOperation.priority = 100;
+                        SceneObject = SceneManager.GetSceneAt(SceneManager.sceneCount - 1);
+                        _steps = ESteps.Checking;
+                    }
+                    else
+                    {
+                        string error = $"Failed to load scene : {MainAssetInfo.AssetPath}";
+                        YooLogger.Error(error);
+                        InvokeCompletion(error, EOperationStatus.Failed);
+                    }
                 }
             }
 
             // 3. 检测加载结果
             if (_steps == ESteps.Checking)
             {
-                Progress = _asyncOperation.progress;
-                if (_asyncOperation.isDone)
+                if (_asyncOperation != null)
                 {
-                    if (SceneObject.IsValid())
+                    if (IsWaitForAsyncComplete || IsForceDestroyComplete)
                     {
-                        InvokeCompletion(string.Empty, EOperationStatus.Succeed);
+                        // 场景加载无法强制异步转同步
+                        YooLogger.Error("The scene is loading asyn !");
                     }
                     else
                     {
-                        string error = $"The load scene is invalid : {MainAssetInfo.AssetPath}";
-                        YooLogger.Error(error);
-                        InvokeCompletion(error, EOperationStatus.Failed);
+                        // 注意：在业务层中途可以取消挂起
+                        if (_asyncOperation.allowSceneActivation == false)
+                        {
+                            if (_suspendLoadMode == false)
+                                _asyncOperation.allowSceneActivation = true;
+                        }
+
+                        Progress = _asyncOperation.progress;
+                        if (_asyncOperation.isDone == false)
+                            return;
                     }
+                }
+
+                if (SceneObject.IsValid())
+                {
+                    InvokeCompletion(string.Empty, EOperationStatus.Succeed);
+                }
+                else
+                {
+                    string error = $"The loaded scene is invalid : {MainAssetInfo.AssetPath}";
+                    YooLogger.Error(error);
+                    InvokeCompletion(error, EOperationStatus.Failed);
                 }
             }
         }
@@ -100,13 +135,12 @@ namespace YooAsset
         /// <summary>
         /// 解除场景加载挂起操作
         /// </summary>
-        public bool UnSuspendLoad()
+        public void UnSuspendLoad()
         {
-            if (_asyncOperation == null)
-                return false;
-
-            _asyncOperation.allowSceneActivation = true;
-            return true;
+            if (IsDone == false)
+            {
+                _suspendLoadMode = false;
+            }
         }
     }
 }