+++ Assets/OxGFrame/AssetLoader/ThirdParty/YooAsset/Runtime/ResourceManager/Provider/DatabaseSceneProvider.cs
@@ -9,14 +9,14 @@ namespace YooAsset
     internal sealed class DatabaseSceneProvider : ProviderBase
     {
         public readonly LoadSceneMode SceneMode;
-        private readonly bool _suspendLoad;
+        private bool _suspendLoadMode;
         private AsyncOperation _asyncOperation;
 
         public DatabaseSceneProvider(ResourceManager manager, string providerGUID, AssetInfo assetInfo, LoadSceneMode sceneMode, bool suspendLoad) : base(manager, providerGUID, assetInfo)
         {
             SceneMode = sceneMode;
             SceneName = Path.GetFileNameWithoutExtension(assetInfo.AssetPath);
-            _suspendLoad = suspendLoad;
+            _suspendLoadMode = suspendLoad;
         }
         internal override void InternalOnStart()
         {
@@ -57,40 +57,66 @@ namespace YooAsset
             // 2. 加载资源对象
             if (_steps == ESteps.Loading)
             {
-                LoadSceneParameters loadSceneParameters = new LoadSceneParameters();
-                loadSceneParameters.loadSceneMode = SceneMode;
-                _asyncOperation = UnityEditor.SceneManagement.EditorSceneManager.LoadSceneAsyncInPlayMode(MainAssetInfo.AssetPath, loadSceneParameters);
-                if (_asyncOperation != null)
+                if (IsWaitForAsyncComplete || IsForceDestroyComplete)
                 {
-                    _asyncOperation.allowSceneActivation = !_suspendLoad;
-                    _asyncOperation.priority = 100;
-                    SceneObject = SceneManager.GetSceneAt(SceneManager.sceneCount - 1);
+                    LoadSceneParameters loadSceneParameters = new LoadSceneParameters(SceneMode);
+                    SceneObject = UnityEditor.SceneManagement.EditorSceneManager.LoadSceneInPlayMode(MainAssetInfo.AssetPath, loadSceneParameters);
                     _steps = ESteps.Checking;
                 }
                 else
                 {
-                    string error = $"Failed to load scene : {MainAssetInfo.AssetPath}";
-                    YooLogger.Error(error);
-                    InvokeCompletion(error, EOperationStatus.Failed);
+                    LoadSceneParameters loadSceneParameters = new LoadSceneParameters(SceneMode);
+                    _asyncOperation = UnityEditor.SceneManagement.EditorSceneManager.LoadSceneAsyncInPlayMode(MainAssetInfo.AssetPath, loadSceneParameters);
+                    if (_asyncOperation != null)
+                    {
+                        _asyncOperation.allowSceneActivation = !_suspendLoadMode;
+                        _asyncOperation.priority = 100;
+                        SceneObject = SceneManager.GetSceneAt(SceneManager.sceneCount - 1);
+                        _steps = ESteps.Checking;
+                    }
+                    else
+                    {
+                        string error = $"Failed to load scene : {MainAssetInfo.AssetPath}";
+                        YooLogger.Error(error);
+                        InvokeCompletion(error, EOperationStatus.Failed);
+                    }
                 }
             }
 
             // 3. 检测加载结果
             if (_steps == ESteps.Checking)
             {
-                Progress = _asyncOperation.progress;
-                if (_asyncOperation.isDone)
+                if (_asyncOperation != null)
                 {
-                    if (SceneObject.IsValid())
+                    if (IsWaitForAsyncComplete || IsForceDestroyComplete)
                     {
-                        InvokeCompletion(string.Empty, EOperationStatus.Succeed);
+                        // 场景加载无法强制异步转同步
+                        YooLogger.Error("The scene is loading asyn !");
                     }
                     else
                     {
-                        string error = $"The loaded scene is invalid : {MainAssetInfo.AssetPath}";
-                        YooLogger.Error(error);
-                        InvokeCompletion(error, EOperationStatus.Failed);
+                        // 注意：在业务层中途可以取消挂起
+                        if (_asyncOperation.allowSceneActivation == false)
+                        {
+                            if (_suspendLoadMode == false)
+                                _asyncOperation.allowSceneActivation = true;
+                        }
+
+                        Progress = _asyncOperation.progress;
+                        if (_asyncOperation.isDone == false)
+                            return;
                     }
+                }
+
+                if (SceneObject.IsValid())
+                {
+                    InvokeCompletion(string.Empty, EOperationStatus.Succeed);
+                }
+                else
+                {
+                    string error = $"The loaded scene is invalid : {MainAssetInfo.AssetPath}";
+                    YooLogger.Error(error);
+                    InvokeCompletion(error, EOperationStatus.Failed);
                 }
             }
 #endif
@@ -99,13 +125,12 @@ namespace YooAsset
         /// <summary>
         /// 解除场景加载挂起操作
         /// </summary>
-        public bool UnSuspendLoad()
+        public void UnSuspendLoad()
         {
-            if (_asyncOperation == null)
-                return false;
-
-            _asyncOperation.allowSceneActivation = true;
-            return true;
+            if (IsDone == false)
+            {
+                _suspendLoadMode = false;
+            }
         }
     }
 }