+++ Assets/OxGFrame/AssetLoader/ThirdParty/YooAsset/Runtime/ResourcePackage/ResourcePackage.cs
@@ -184,7 +184,8 @@ namespace YooAsset
                 var initializeParameters = parameters as WebPlayModeParameters;
                 initializeOperation = webPlayModeImpl.InitializeAsync(assist,
                     initializeParameters.BuildinQueryServices,
-                    initializeParameters.RemoteServices);
+                    initializeParameters.RemoteServices,
+                    initializeParameters.WechatQueryServices);
             }
             else
             {
@@ -413,10 +414,13 @@ namespace YooAsset
             return operation;
         }
 
+        /// <summary>
+        /// 获取指定版本的缓存信息
+        /// </summary>
         public GetAllCacheFileInfosOperation GetAllCacheFileInfosAsync(string packageVersion)
         {
             DebugCheckInitialize();
-            
+
             var operation = new GetAllCacheFileInfosOperation(_persistentMgr, _cacheMgr, packageVersion);
             OperationSystem.StartOperation(PackageName, operation);
             return operation;
@@ -600,6 +604,29 @@ namespace YooAsset
 
         #region 场景加载
         /// <summary>
+        /// 同步加载场景
+        /// </summary>
+        /// <param name="location">场景的定位地址</param>
+        /// <param name="sceneMode">场景加载模式</param>
+        public SceneHandle LoadSceneSync(string location, LoadSceneMode sceneMode = LoadSceneMode.Single)
+        {
+            DebugCheckInitialize();
+            AssetInfo assetInfo = ConvertLocationToAssetInfo(location, null);
+            return LoadSceneInternal(assetInfo, true, sceneMode, false, 0);
+        }
+
+        /// <summary>
+        /// 同步加载场景
+        /// </summary>
+        /// <param name="assetInfo">场景的资源信息</param>
+        /// <param name="sceneMode">场景加载模式</param>
+        public SceneHandle LoadSceneSync(AssetInfo assetInfo, LoadSceneMode sceneMode = LoadSceneMode.Single)
+        {
+            DebugCheckInitialize();
+            return LoadSceneInternal(assetInfo, true, sceneMode, false, 0);
+        }
+
+        /// <summary>
         /// 异步加载场景
         /// </summary>
         /// <param name="location">场景的定位地址</param>
@@ -610,8 +637,7 @@ namespace YooAsset
         {
             DebugCheckInitialize();
             AssetInfo assetInfo = ConvertLocationToAssetInfo(location, null);
-            var handle = _resourceMgr.LoadSceneAsync(assetInfo, sceneMode, suspendLoad, priority);
-            return handle;
+            return LoadSceneInternal(assetInfo, false, sceneMode, suspendLoad, priority);
         }
 
         /// <summary>
@@ -624,7 +650,16 @@ namespace YooAsset
         public SceneHandle LoadSceneAsync(AssetInfo assetInfo, LoadSceneMode sceneMode = LoadSceneMode.Single, bool suspendLoad = false, uint priority = 0)
         {
             DebugCheckInitialize();
+            return LoadSceneInternal(assetInfo, false, sceneMode, suspendLoad, priority);
+        }
+
+        private SceneHandle LoadSceneInternal(AssetInfo assetInfo, bool waitForAsyncComplete, LoadSceneMode sceneMode, bool suspendLoad, uint priority)
+        {
+            DebugCheckAssetLoadMethod(nameof(LoadAssetAsync));
+            DebugCheckAssetLoadType(assetInfo.AssetType);
             var handle = _resourceMgr.LoadSceneAsync(assetInfo, sceneMode, suspendLoad, priority);
+            if (waitForAsyncComplete)
+                handle.WaitForAsyncComplete();
             return handle;
         }
         #endregion