+++ Assets/OxGFrame/AssetLoader/ThirdParty/YooAsset/Runtime/Utility/YooUtility.cs
@@ -116,6 +116,18 @@ namespace YooAsset
     internal static class FileUtility
     {
         /// <summary>
+        /// 字节数据转换成文本数据
+        /// </summary>
+        /// <param name="data"></param>
+        /// <returns></returns>
+        public static string BytesToText(byte[] data)
+        {
+            UTF8Encoding utf8 = new UTF8Encoding();
+            string txt = utf8.GetString(data);
+            return txt;
+        }
+
+        /// <summary>
         /// 读取文件的文本数据
         /// </summary>
         public static string ReadAllText(string filePath)
@@ -215,12 +227,20 @@ namespace YooAsset
         /// </summary>
         public static string FileSHA1(string filePath)
         {
+            using (FileStream fs = new FileStream(filePath, FileMode.Open, FileAccess.Read, FileShare.Read))
+            {
+                return StreamSHA1(fs);
+            }
+        }
+
+        /// <summary>
+        /// 获取文件的Hash值
+        /// </summary>
+        public static string FileSHA1Safely(string filePath)
+        {
             try
             {
-                using (FileStream fs = new FileStream(filePath, FileMode.Open, FileAccess.Read, FileShare.Read))
-                {
-                    return StreamSHA1(fs);
-                }
+                return FileSHA1(filePath);
             }
             catch (Exception e)
             {
@@ -267,12 +287,20 @@ namespace YooAsset
         /// </summary>
         public static string FileMD5(string filePath)
         {
+            using (FileStream fs = new FileStream(filePath, FileMode.Open, FileAccess.Read, FileShare.Read))
+            {
+                return StreamMD5(fs);
+            }
+        }
+
+        /// <summary>
+        /// 获取文件的MD5
+        /// </summary>
+        public static string FileMD5Safely(string filePath)
+        {
             try
             {
-                using (FileStream fs = new FileStream(filePath, FileMode.Open, FileAccess.Read, FileShare.Read))
-                {
-                    return StreamMD5(fs);
-                }
+                return FileMD5(filePath);
             }
             catch (Exception e)
             {
@@ -317,12 +345,20 @@ namespace YooAsset
         /// </summary>
         public static string FileCRC32(string filePath)
         {
+            using (FileStream fs = new FileStream(filePath, FileMode.Open, FileAccess.Read, FileShare.Read))
+            {
+                return StreamCRC32(fs);
+            }
+        }
+
+        /// <summary>
+        /// 获取文件的CRC32
+        /// </summary>
+        public static string FileCRC32Safely(string filePath)
+        {
             try
             {
-                using (FileStream fs = new FileStream(filePath, FileMode.Open, FileAccess.Read, FileShare.Read))
-                {
-                    return StreamCRC32(fs);
-                }
+                return FileCRC32(filePath);
             }
             catch (Exception e)
             {