+++ Assets/OxGFrame/Hotfixer/Scripts/Runtime/Hotfixers.cs
@@ -1,3 +1,4 @@
+﻿using OxGFrame.AssetLoader.Bundle;
 using System.Reflection;
 
 namespace OxGFrame.Hotfixer
@@ -5,7 +6,7 @@ namespace OxGFrame.Hotfixer
     public static class Hotfixers
     {
         /// <summary>
-        /// Start hotfix files download and load all
+        /// Start hotfix files download and load all (default is AppPackageInfoWithBuild)
         /// </summary>
         /// <param name="packageName"></param>
         /// <param name="aotAssemblies"></param>
@@ -13,6 +14,17 @@ namespace OxGFrame.Hotfixer
         public static void CheckHotfix(string packageName, string[] aotAssemblies, string[] hotfixAssemblies)
         {
             HotfixManager.GetInstance().CheckHotfix(packageName, aotAssemblies, hotfixAssemblies);
+        }
+
+        /// <summary>
+        /// Start hotfix files download and load all
+        /// </summary>
+        /// <param name="packageInfoWithBuild"></param>
+        /// <param name="aotAssemblies"></param>
+        /// <param name="hotfixAssemblies"></param>
+        public static void CheckHotfix(PackageInfoWithBuild packageInfoWithBuild, string[] aotAssemblies, string[] hotfixAssemblies)
+        {
+            HotfixManager.GetInstance().CheckHotfix(packageInfoWithBuild, aotAssemblies, hotfixAssemblies);
         }
 
         /// <summary>